# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/config": {
            "get": {
                "description": "获取用户配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "获取用户配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": ".更改用户配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": ".更改用户配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户配置信息",
                        "name": "configuration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.Configuration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/configState": {
            "get": {
                "description": ".获取状态配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": ".获取状态配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": ".更改状态配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": ".更改状态配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "With the bearer started JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "状态配置信息",
                        "name": "configState",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/api.ConfigState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Error\":\"Success\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "301": {
                        "description": "{\"Error\":\"Re-login\",\"Data\": object}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"Error\":\"error\",\"Data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ConfigState": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "states": {
                    "type": "string"
                }
            }
        },
        "api.Configuration": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "用户ID --不填",
                    "type": "integer"
                },
                "language": {
                    "description": "语言选择 1简体中文 2繁体中文 3英文 不更改不填或填0",
                    "type": "integer"
                },
                "message": {
                    "description": "消息推送开关 1开2关 不更改不填或填0",
                    "type": "integer"
                },
                "sound": {
                    "description": "声音开关 1开2关 不更改不填或填0",
                    "type": "integer"
                }
            }
        },
...
    }
}